+++ tools/gfx/d3d11/d3d11-shader-program.cpp
@@ -1,0 +1,17 @@
+// d3d11-shader-program.cpp
+#include "d3d11-shader-program.h"
+
+#include "d3d11-base.h"
+
+namespace gfx
+{
+
+using namespace Slang;
+
+namespace d3d11
+{
+
+
+
+} // namespace d3d11
+} // namespace gfx
